services/parser/requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
lxml>=5.3.0
redis>=5.2.0
apscheduler>=3.10.0
geopy>=2.4.0
python-dotenv>=1.0.0
pydantic>=2.10.0
httpx>=0.28.0
